requirements.txt
pyTelegramBotAPI==4.14.0
requests==2.32.5
python-dotenv==1.0.1
